requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3
